Day80
Alice and Bob went to a pet store. There are N animals in the store where the ith animal is of type Ai?.

Alice decides to buy some of these N animals. Bob decides that he will buy all the animals left in the store after Alice has made the purchase.

Find out whether it is possible that Alice and Bob end up with exactly same multiset of animals.